Fix_numpy_compatibility

+++ text2interaction/common/quaternion.py
@@ -11,7 +11,7 @@ import torch.nn.functional as F
 
 _EPS4 = np.finfo(float).eps * 4.0
 
-_FLOAT_EPS = np.finfo(np.float).eps
+_FLOAT_EPS = np.finfo(float).eps
 
 # PyTorch-backed implementations
 def qinv(q):
